requirements.txt
gensim==3.8.3
pandas==1.1.3
numpy==1.18.5
scikit-learn==0.23.2
streamlit==0.68.0
st-annotated-text==1.0.1
nltk=3.5
